Cleaned up faq code, created ask question button component, added button globally
