+++ client/src/App.jsx
@@ -7,9 +7,11 @@ import { Navbar } from './components/Navbar/Navbar';
 import { Footer } from './components/footer/Footer';
 import { useDispatch, useSelector } from 'react-redux';
 import { initialsSelector, loggedInSelector, registeredSelector } from './pages/userSlice';
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { getUserInfo } from './pages/Login/saga';
 import { InitialPage } from './pages/Initial/Initial';
+import { AskQuestionButton } from './components/button/AskQuestionButton/AskQuestionButton';
+import { userSelector } from '../src/pages/userSlice';
 
 export default function App() {
   const dispatch = useDispatch();
@@ -30,10 +32,19 @@ const TransitionRoutes = () => {
   const loggedIn = useSelector(loggedInSelector);
   const registered = useSelector(registeredSelector);
   const initials = useSelector(initialsSelector);
+  const [pageState, setPageState] = useState('form');
+  const [showPopUp, setShowPopUp] = useState(false);
+  const { user } = useSelector(userSelector);
   console.log(loggedIn);
   return (
     <TransitionGroup>
-      <Navbar isLoggedIn={loggedIn} froshInitials={initials} isRegistered={registered} />
+      <div
+        onClick={() => {
+          setShowPopUp(false);
+        }}
+      >
+        <Navbar isLoggedIn={loggedIn} froshInitials={initials} isRegistered={registered} />
+      </div>
       <ScrollToTop />
       <CSSTransition key={location.key} classNames="page" timeout={300}>
         <Routes location={location}>
@@ -54,6 +65,13 @@ const TransitionRoutes = () => {
           <Route path="*" element={pages['404'].component} />
         </Routes>
       </CSSTransition>
+      <AskQuestionButton
+        showPopUp={showPopUp}
+        setShowPopUp={setShowPopUp}
+        pageState={pageState}
+        setPageState={setPageState}
+        user={user}
+      />
     </TransitionGroup>
   );
 };
