+++ client/src/pages/FAQ/FAQ.jsx
@@ -1,22 +1,12 @@
-import { React, useState, useEffect, useRef } from 'react';
+import { React, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { getQuestions, submitQuestion } from './functions';
+import { getQuestions } from './functions';
 import './FAQ.scss';
 import Wave from '../../assets/misc/wave-reverse.png';
 import { ButtonSelector } from '../../components/buttonSelector/buttonSelector/ButtonSelector';
 import { SingleAccordion } from '../../components/text/Accordion/SingleAccordion/SingleAccordion';
-import { Button } from '../../components/button/Button/Button';
-import { ButtonOutlined } from '../../components/button/ButtonOutlined/ButtonOutlined';
-import { TextInput } from '../../components/input/TextInput/TextInput';
 import SearchIcon from '../../assets/misc/magnifying-glass-solid.svg';
 import DeleteIcon from '../../assets/misc/circle-xmark-solid.svg';
-import { ErrorSuccessBox } from '../../components/containers/ErrorSuccessBox/ErrorSuccessBox';
-import LoadingAnimation from '../../components/misc/LoadingAnimation/LoadingAnimation';
-import QuestionMark from '../../../assets/icons/question-mark-solid.svg';
-
-import { PopupModal } from '../../components/popup/PopupModal';
-import { userSelector } from '../userSlice';
-import { useSelector } from 'react-redux';
 
 const PageFAQ = () => {
   const [activeIndex, setActiveIndex] = useState(0);
@@ -24,15 +14,12 @@ const PageFAQ = () => {
   const [isMultiSearch, setIsMultiSearch] = useState(false);
   const [isNoMatch, setIsNoMatch] = useState(false);
   const [selectedSearchResult, setSelectedSearchResult] = useState(false);
-  const [selectedQuestion, setSelectedQuestion] = useState(0);
   const [selectedQuestions, setSelectedQuestions] = useState([]);
   const [searchQuery, setSearchQuery] = useState('');
-  const [pageState, setPageState] = useState('form');
   const data = getQuestions();
   const unsortedQuestions = [];
   const questionsObjects = {};
   const questionCategories = [];
-  const { user } = useSelector(userSelector);
   for (let i = 0; i < data.length; i++) {
     if (!questionsObjects.hasOwnProperty(data[i].category)) {
       questionsObjects[data[i].category] = [];
@@ -66,7 +53,6 @@ const PageFAQ = () => {
           questions={unsortedQuestions}
           setIsSearch={setIsSearch}
           setIsMultiSearch={setIsMultiSearch}
-          setSelectedQuestion={setSelectedQuestion}
           setSelectedQuestions={setSelectedQuestions}
           searchQuery={searchQuery}
           setSearchQuery={setSearchQuery}
@@ -109,16 +95,6 @@ const PageFAQ = () => {
         >
           <div
             className={`${
-              !isMultiSearch && !isNoMatch ? 'faq-show-accordion' : 'faq-hide-accordion'
-            }`}
-          >
-            <FAQDisplaySearchQuestion
-              selectedQuestion={selectedQuestion}
-              questions={unsortedQuestions}
-            />
-          </div>
-          <div
-            className={`${
               isMultiSearch && !isNoMatch ? 'faq-show-accordion' : 'faq-hide-accordion'
             }`}
           >
@@ -131,7 +107,6 @@ const PageFAQ = () => {
             <h1>No results</h1>
           </div>
         </div>
-        <FAQFab pageState={pageState} setPageState={setPageState} user={user} />
       </div>
     </div>
   );
@@ -140,7 +115,6 @@ const PageFAQ = () => {
 const FAQPageHeader = ({
   questions,
   setIsSearch,
-  setSelectedQuestion,
   searchQuery,
   setSearchQuery,
   selectedSearchResult,
@@ -184,7 +158,6 @@ const FAQPageHeader = ({
                     ? '15px 0px 0px 0px'
                     : '',
               }}
-              // onClick={() => handleSearchIconClick()}
             >
               <img src={SearchIcon} alt="Search Button" height={30} />
             </div>
@@ -194,7 +167,6 @@ const FAQPageHeader = ({
                 setSearchQuery={setSearchQuery}
                 setIsSearch={setIsSearch}
                 setIsMultiSearch={setIsMultiSearch}
-                setSelectedQuestion={setSelectedQuestion}
                 questions={filteredQuestions}
                 selectedSearchResult={selectedSearchResult}
                 setSelectedSearchResult={setSelectedSearchResult}
@@ -234,7 +206,6 @@ const FAQPageHeader = ({
 FAQPageHeader.propTypes = {
   questions: PropTypes.array.isRequired,
   setIsSearch: PropTypes.func.isRequired,
-  setSelectedQuestion: PropTypes.func.isRequired,
   searchQuery: PropTypes.string.isRequired,
   setSearchQuery: PropTypes.func.isRequired,
   selectedSearchResult: PropTypes.number.isRequired,
@@ -320,14 +291,10 @@ const FAQSearchBar = ({
   setSearchQuery,
   setIsSearch,
   setIsMultiSearch,
-  setSelectedQuestion,
   questions,
-  selectedSearchResult,
   setSelectedSearchResult,
   setIsNoMatch,
   setSelectedQuestions,
-  setActiveIndex,
-  questionCategories,
 }) => {
   useEffect(() => {
     setIsSearch(true);
@@ -349,7 +316,6 @@ const FAQSearchBar = ({
       setIsSearch(false);
     }
   };
-
   return (
     <div className={'faq-search'}>
       <div className={'faq-search-input'}>
@@ -369,28 +335,11 @@ FAQSearchBar.propTypes = {
   setSearchQuery: PropTypes.func.isRequired,
   setIsSearch: PropTypes.func.isRequired,
   setIsMultiSearch: PropTypes.func.isRequired,
-  setSelectedQuestion: PropTypes.func.isRequired,
   questions: PropTypes.array.isRequired,
   selectedSearchResult: PropTypes.bool.isRequired,
   setSelectedSearchResult: PropTypes.func.isRequired,
   setIsNoMatch: PropTypes.bool.isRequired,
   setSelectedQuestions: PropTypes.func.isRequired,
-  setActiveIndex: PropTypes.number.isRequired,
-  questionCategories: PropTypes.array.isRequired,
-};
-
-const FAQDisplaySearchQuestion = ({ selectedQuestion, questions }) => {
-  return (
-    <div className={'faq-search-result-container'}>
-      <div className={'faq-search-result-question'}>{questions[selectedQuestion].question}</div>
-      <div className={'faq-search-result-answer'}>{questions[selectedQuestion].answer}</div>
-    </div>
-  );
-};
-
-FAQDisplaySearchQuestion.propTypes = {
-  selectedQuestion: PropTypes.number.isRequired,
-  questions: PropTypes.array.isRequired,
 };
 
 const FAQDisplayAllSearchQuestion = ({ selectedQuestions, questions }) => {
@@ -406,159 +355,6 @@ const FAQDisplayAllSearchQuestion = ({ selectedQuestions, questions }) => {
 FAQDisplayAllSearchQuestion.propTypes = {
   selectedQuestions: PropTypes.array.isRequired,
   questions: PropTypes.array.isRequired,
-};
-
-const FAQAskQuestion = ({ pageState, setPageState, user }) => {
-  const [signUpError, setSignUpError] = useState('');
-  const [errorColor, setErrorColor] = useState(false);
-  const initialFormData = {
-    question: '',
-    email: user?.email ?? '',
-  };
-  const [emailText, setEmailText] = useState({ email: user ? user.email : '' });
-  const [questionText, setQuestionText] = useState({});
-  const [formData, updateFormData] = useState(initialFormData);
-  const [clearText, setClearText] = useState(false);
-  useEffect(() => {
-    updateFormData({ question: questionText.question, email: emailText.email });
-  }, [emailText, questionText]);
-  const handleChangeEmail = (text) => {
-    let newEmailState = { ...emailText };
-    newEmailState.email = text;
-    setEmailText(newEmailState);
-  };
-  const handleChangeQuestion = (text) => {
-    let newQuestionState = { ...questionText };
-    newQuestionState.question = text;
-    setQuestionText(newQuestionState);
-  };
-  async function handleSubmit(text) {
-    // console.log(submitQuestion(formData));
-    //console.log(formData);
-    if (formData?.question?.length > 0 && formData?.email?.length > 0) {
-      setPageState('loading');
-      const result = await submitQuestion(formData);
-      if (result !== true) {
-        setSignUpError(result);
-        setErrorColor(true);
-        setPageState('form');
-      } else {
-        setPageState('success');
-        updateFormData(initialFormData);
-        setClearText(true);
-        setSignUpError('Thank you for submitting your question!');
-        setErrorColor(false);
-        setPageState('form');
-      }
-    }
-  }
-
-  return (
-    <div className={'faq-ask-question-container'}>
-      <div className={'faq-ask-question-stacking'}>
-        <div
-          className={`faq-page-questions ${
-            pageState !== 'form' ? 'faq-page-disappear' : 'faq-page-appear'
-          }`}
-        >
-          <h1 className={'faq-ask-question-title'}>Can&apos;t Find Your Question?</h1>
-          <ErrorSuccessBox content={signUpError} error={errorColor} success={!errorColor} />
-          <form>
-            <label>
-              <div className={'faq-ask-question-email-box'}>
-                <TextInput
-                  onChange={(text) => handleChangeEmail(text)}
-                  inputType={'text'}
-                  placeholder={'Email'}
-                  initialValue={emailText.email}
-                  style={{ height: '45px' }}
-                  clearText={clearText}
-                  setClearText={setClearText}
-                />
-              </div>
-            </label>
-            <label>
-              <div className={'faq-ask-question-box'}>
-                <TextInput
-                  onChange={(text) => handleChangeQuestion(text)}
-                  inputType={'textArea'}
-                  placeholder={'Type your question here...'}
-                  style={{ height: '150px', resize: 'vertical' }}
-                  clearText={clearText}
-                  setClearText={setClearText}
-                />
-              </div>
-            </label>
-            <div style={{ textAlign: 'center' }}>
-              <Button label={'Submit'} onClick={handleSubmit}>
-                Submit
-              </Button>
-            </div>
-          </form>
-        </div>
-        <div
-          className={`faq-loading ${pageState === 'loading' ? 'faq-loading-appear' : ''} 
-          ${pageState === 'success' ? 'faq-loading-disappear' : ''}`}
-        >
-          <LoadingAnimation size={'60px'} />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-FAQAskQuestion.propTypes = {
-  pageState: PropTypes.string.isRequired,
-  setPageState: PropTypes.string.isRequired,
-  user: PropTypes.object,
-};
-
-const FAQFab = ({ pageState, setPageState, user }) => {
-  const [showPopUp, setShowPopUp] = useState(false);
-
-  return (
-    <>
-      <PopupModal
-        trigger={showPopUp}
-        setTrigger={setShowPopUp}
-        blurBackground={false}
-        exitIcon={true}
-      >
-        <div className="ask-question-popup">
-          <div className={'faq-ask-question-outer-container'}>
-            <FAQAskQuestion pageState={pageState} setPageState={setPageState} user={user} />
-          </div>
-        </div>
-      </PopupModal>
-      <div className={'faq-fab'}>
-        <Button
-          style={{ boxShadow: '5px 5px 20px #13131362' }}
-          class_options={'faq-fab-button'}
-          label={
-            <div className={'faq-fab-container'}>
-              <img
-                className={'faq-fab-icon'}
-                src={QuestionMark}
-                alt="Question Button"
-                height={30}
-              />
-              <span className={'faq-fab-content'}>Ask a Question</span>
-            </div>
-          }
-          isSecondary
-          onClick={() => {
-            setShowPopUp(true);
-          }}
-        />
-      </div>
-    </>
-  );
-};
-
-FAQFab.propTypes = {
-  pageState: PropTypes.string.isRequired,
-  setPageState: PropTypes.string.isRequired,
-  user: PropTypes.object,
 };
 
 export { PageFAQ };
