+++ client/src/pages/FAQ/FAQ.scss
@@ -269,123 +269,7 @@
   padding-bottom: 5px;
 }
 
-.faq-ask-question-outer-container {
-  width: 100%;
-  margin-left: auto;
-  margin-right: auto;
-  @include devices(tablet) {
-    width: 100%;
-  }
-}
-
-.faq-ask-question-container {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.faq-ask-question-stacking {
-  display: grid;
-}
-
-.faq-page-questions {
-  grid-column: 1;
-  grid-row: 1;
-}
-
-.faq-ask-question-title {
-  text-align: center;
-  width: 100%;
-}
-
-.faq-ask-question-box {
-  margin: 20px 0px 0px 0px;
-  text-align: center;
-}
-
 .faq-no-results {
   width: 100%;
   text-align: center;
-}
-
-.faq-loading {
-  z-index: -1;
-  transition: transform 400ms, opacity 400ms;
-  opacity: 0;
-  transform: scale(0.5);
-  grid-column: 1;
-  grid-row: 1;
-}
-
-.faq-loading-appear {
-  transition: transform 400ms, opacity 400ms;
-  opacity: 1;
-  transform: scale(1);
-}
-
-.faq-loading-disappear {
-  transition: transform 1500ms, opacity 900ms;
-  opacity: 0;
-  transform: scale(10);
-}
-
-.faq-page-disappear {
-  transform: translateY(-5vh);
-  opacity: 0;
-  transition: transform 1200ms, opacity 1000ms;
-  user-select: none;
-  pointer-events: none;
-}
-
-.faq-page-appear {
-  opacity: 1;
-  transition: transform 1200ms, opacity 1000ms;
-  page-disappear {
-    transform: translateY(-35vh);
-    opacity: 0;
-    transition: transform 1200ms, opacity 1200ms;
-    user-select: none;
-    pointer-events: none;
-  }
-}
-
-.faq-ask-question-email-box {
-  margin: 20px 0px 0px 0px;
-  text-align: center;
-}
-
-.faq-fab {
-  bottom: 8px;
-  right: 10px;
-  position: fixed;
-  z-index: 5;
-  @include devices(tablet) {
-    bottom: 50px;
-    right: 5px;
-  }
-}
-
-.faq-fab-button {
-  @include devices(tablet) {
-    padding: 10px 15px !important;
-  }
-}
-
-.faq-fab-container {
-  display: flex;
-  align-items: center;
-}
-
-.faq-fab-icon {
-  margin-right: 15px;
-}
-
-.ask-question-popup {
-  padding: 0px 50px;
-  h1 {
-    color: var(--white);
-  }
-  @include devices(tablet) {
-    padding: 0px 20px;
-  }
 }
